CaretexBackend/requirements.txt
appdirs==1.4.4
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
beautifulsoup4==4.9.3
certifi==2021.5.30
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-bootstrap4==3.0.1
django-heroku==0.3.1
django-import-export==2.5.0
django-multiselectfield==0.1.12
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
six==1.15.0
soupsieve==2.2.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tablib==3.0.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
virtualenv==20.4.4
whitenoise==5.3.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
